refactored code and setup calculation for total return up to each date NOTE:values are still in decimal format and are not denoted as percentages
