+++ investure_app/server/index.js
@@ -1,7 +1,9 @@
 const express = require("express")
-const xlsx = require("xlsx")
 
+const xlsx = require("xlsx")
 const fileUpload = require("express-fileupload")
+const processFileUpload = require("./filehandler")
+const calculateTotalReturns = require("./calculations")
 
 const PORT = process.env.PORT || 5001;
 
@@ -10,33 +12,13 @@ const app = express()
 const inMemoryDataStore = {}
 
 
-
-
 //Middleware to convert file into computer readable data
-app.use(fileUpload({ limits: { fileSize: 10 * 1024 * 1024 } }));
+app.use(fileUpload({ limits: { fileSize: 100 * 10024 * 10024 } }));
 
 app.post('/upload', (req,res) => {
-
     try {
         if(req.files && req.files.file) {
-            const uploadFile = req.files.file
-
-            const workbook = xlsx.read(uploadFile.data, {type: "buffer"})
-
-            const sheetNames = workbook.SheetNames[0]
-            const rawdata_worksheet = workbook.Sheets[sheetNames]
-
-
-            const jsonData = xlsx.utils.sheet_to_json(rawdata_worksheet);
-
-            jsonData.forEach(row => {
-                const dateKey = row.ReferenceDate
-                const value = row.DailyReturn
-                if (dateKey) {
-                    inMemoryDataStore[dateKey] = value
-                }
-            })
-
+            processFileUpload(req.files.file, inMemoryDataStore)
             res.status(200).json({message: 'File uploaded and data stored successfully'});
 
         }
@@ -51,7 +33,8 @@ app.post('/upload', (req,res) => {
 app.get('/getTotalReturns', (req, res) => {
     try{
         if(Object.keys(inMemoryDataStore).length > 0) {
-            res.status(200).json({message: "endpoint is setup"})
+            const cumulative = calculateTotalReturns(inMemoryDataStore)
+            res.status(200).json(cumulative)
         }
         else{
             res.status(400).json({message: "There is no data to fetch, please make sure you have uploaded a file"})
